Mendeleev.jl

A package for accessing chemical elements data. It's code was initially based on PeriodicTable Julia package, whereas the data come mainly from the Python mendeleev package.

Installation

As usual

] add Mendeleev

Intentions

The aim was to have a package compatible with PeriodicTable as possible, but with much more comprehensive data from possibly reliable and traceable sources. The package should still be possibly lightweight like PeriodicTable. Mendeleev, like it's predessor, has only one direct dependence (Unitful), keeps data in static form, and takes about the same time to load (about half a second on a medium-range medium-aged notebook). See below for compatibility details.

Another aim was to make it easier to update data. Most data are read out from the (Python) mendeleev database by a separate external skript and converted to static Julia code, which further can be relatively straightforwardly edited, should that be necessary.

Usage

Mendeleev.jl provides a Julia interface to a database of element properties for all of the elements in the periodic table. In particular Mendeleev exports a global variable called chem_elements, which is a collection of ChemElem data structures.

julia> using Mendeleev

julia> chem_elements
Elements(…118 elements…):
H                                                  He
Li Be                               B  C  N  O  F  Ne
Na Mg                               Al Si P  S  Cl Ar
K  Ca Sc Ti V  Cr Mn Fe Co Ni Cu Zn Ga Ge As Se Br Kr
Rb Sr Y  Zr Nb Mo Tc Ru Rh Pd Ag Cd In Sn Sb Te I  Xe
Cs Ba    Hf Ta W  Re Os Ir Pt Au Hg Tl Pb Bi Po At Rn
Fr Ra    Rf Db Sg Bh Hs Mt Ds Rg Cn Nh Fl Mc Lv Ts Og
Uue                                                   
      La Ce Pr Nd Pm Sm Eu Gd Tb Dy Ho Er Tm Yb Lu    
      Ac Th Pa U  Np Pu Am Cm Bk Cf Es Fm Md No Lr

You can look up elements by name (case-insensitive) via chem_elements["oxygen"], by symbol via chem_elements[:O], or by number via chem_elements[8], for example.

All physical quantities are unitful.

The data is pretty-printed when you look up an element in the Julia REPL. For example:

julia> chem_elements["oxygen"]
Oxygen (O), number 8:
        category: diatomic nonmetal
     atomic mass: 15.999 u
natural isotopes: ((99.738% ¹⁶O m=15.99491462 u ), (0.04% ¹⁷O m=16.999131757 u ), (0.222% ¹⁸O m=17.999159613 u ))
         density: 0.001308 g/cm³
 molar heat cap.: 29.378 J/mol⋅K
   melting point: 54.8 K
   boiling point: 90.19 K
           phase: Gas
          shells: [2, 6]
e⁻-configuration: 1s² 2s² 2p⁴
         summary: Oxygen is a chemical element with symbol O and atomic number 8. It is a member of the chalcogen group on the periodic table and is a highly reactive nonmetal and oxidizing agent that readily forms compounds (notably oxides) with most elements. By mass, oxygen is the third-most abundant element in the universe, after hydrogen and helium.
  CAS identifier: 7782-44-7
   discovered by: 1774 by Joseph Priestly, Carl Wilhelm Scheele in England/Sweden
    NIST webbook: https://webbook.nist.gov/cgi/inchi/InChI%3D1S/O2/c1-2
   wikipedia URL: https://en.wikipedia.org/wiki/Oxygen
  spectral image: https://en.wikipedia.org/wiki/File:Oxygen_spectre.jpg

Alternatively, you may want to get a list of elements,

julia> chem_elements[1:4]
4-element Vector{ChemElem}:
    Element(Hydrogen)
    Element(Helium)
    Element(Lithium)
    Element(Beryllium)

Data by

The data used for this package has been pulled up mainly from mendeleev by Lukasz Mentel. See mendeleev documentation for the data sources. Some information (but no physical quantities) taken over from PeriodicTable.jl, which was in turn taken mostly from here.

List of ChemElement properties:

Elements Data Fields

Types and Functions

Mendeleev.ChemElem — Type

ChemElem

This struct's fields correspond to only the most basic data pertaining to a chemical element, all other data being accessed via properties. It is exported.

    atomic_number::Int
    name::String
    symbol::Symbol

Mendeleev.ChemElems — Type

ChemElems

This struct is a container for elements and provides access by number, symbol, or name. It is not exported.

Mendeleev.Group_M — Type

Group_M

This struct describes a group of the periodic table. It is not exported.

    no::Int
    symbol::String
    name::String

Mendeleev.Isotope — Type

Isotope

This struct describes a single isotope. Only naturally occuring stable or almost stable isotopes are included. It is not exported.

    atomic_number::Int
    mass_number::Int
    mass::typeof(1.0*u"u")
    abundance::Float64

Mendeleev.Isotopes — Type

Isotopes

This struct is a container for isotopes of an element. It provides access only by position(s) in the array of the (stable or almost stable) isotopes of this elements. More comprehensive isotopes data to be published in a separate package. It is not exported.

Mendeleev.ScreenConst — Type

ScreenConst

This struct describes screening constant for a given orbital of a given element. It is not exported.

    atomic_number::Int
    shell::Int
    orb_type::Int
    screening::Float64

Mendeleev.ScreenConstants — Type

ScreenConstants

This struct is a container for screening constants of an element. See examples below for indexing access. It is not exported.

Examples

julia> chem_elements[:Mo].sconst[2, 1]
Mo 2s: 11.1232

julia> chem_elements[:Mo].sconst[2, 's']
Mo 2s: 11.1232

julia> chem_elements[:Mo].sconst[2, "s"]
Mo 2s: 11.1232

julia> chem_elements[:Mo].sconst["2s"]
Mo 2s: 11.1232

Mendeleev.fn_discovered_by — Method

fn_discovered_by(e::ChemElem)

Returns a string with information (if available) about the year of discovery, names of discoverers, and discovery location

Mendeleev.fn_mass_number — Method

fn_mass_number(e::ChemElem)

Return the mass number of the most abundant isotope, if there are any stable isotopes for this element. Otherwise returns missing

Mendeleev.fn_neutrons — Method

fn_neutrons(e::ChemElem) Return the number of neutrons for the most abundant isotope, if there are any stable isotopes for this element. Otherwise returns missing

Compatibility Issues

PeriodicTable.jl exports global variable called elements, which is a collection of Element data structures. In Mendeleev.jl these called correspondingly chem_elements and ChemElem, the word "element" having too much meanings IMO. For legacy application you however still may use the old names:

julia> using Mendeleev: elements, Element

One property (source) got a different meaning, one (color) was omitted completely, and two (discovered_by, named_by) are omitted, but there exist similar properties. For details, see Elements Data section. Missing data were encoded in the PeriodicTable.jl resp. by empty strings and by NaN, as the package predates introduction of Missing. Here, we using Missing.

Developed by

Eben60

Credits

Developers of PeriodicTable.jl

Rahul Lakhanpal

Steven G. Johnson

Jacob Wikmark

Carsten Bauer

Some data originally from

here

Most data from mendeleev by Lukasz Mentel

Wikipedia and CRC Handbook of Chemistry and Physics (ISBN 9781482208689) were the most used primary data sources for words, words, words and for numbers, respectively.