Types

Mendeleev.ChemElem — Type

ChemElem

This struct's fields correspond to only the most basic data pertaining to a chemical element, all other data being accessed via properties. It is exported.

    atomic_number::Int
    name::String
    symbol::Symbol

Mendeleev.ChemElems — Type

ChemElems

This struct is a container for elements and provides access by number, symbol, or name. It is not exported.

Mendeleev.Electronegativities — Type

Electronegativities

This struct contains electronegativities of an element according to different scales. Any field except atomic_number can also be missing, the representation below is simplified for legibility. It is not exported.

    Allen::typeof(1.0*u"eV")
    Allred::typeof(1.0(u"e_au"^2/u"pm"^2))
    Cottrell::typeof(1.0(u"e_au"/u"pm")^(1//2))
    Ghosh::Float64 # dimensionality to be clarified yet
    Gordy::typeof(1.0(u"e_au"/u"pm"))
    Martynov::typeof(1.0*u"eV^(1//2)")
    Mulliken::typeof(1.0*u"eV")
    Nagle::Float64 # dimensionality to be clarified yet
    Pauling::typeof(1.0*u"eV^(1//2)")
    Sanderson::Float64
    Li::LiXue

Mendeleev.Group_M — Type

Group_M

This struct describes a group of the periodic table. It is not exported.

    no::Int
    symbol::String
    name::String

Mendeleev.IonicRadii — Method

IonicRadii

This struct is a container for ionic radii of an element. It provides access by position(s) in the array of IonicRadius structs for the given element, as well as filtering according to (; charge, coordination, spin, econf, most_reliable). It is not exported.

Examples

julia> feir = chem_elements.Fe.ionic_radii;

julia> feir[2]
(Fe2+, coordination=IVSQ, econf=3d6, spin=HS, crystal_radius=78.0 pm, ionic_radius=64.0 pm, ionic_potential=0.03125 e pm⁻¹, most_reliable=false)

julia> feir[[2,3]]
2-element Vector{Mendeleev.IonicRadius}:
 (Fe2+, coordination=IVSQ, econf=3d6, spin=HS, crystal_radius=78.0 pm, ionic_radius=64.0 pm, ionic_potential=0.03125 e pm⁻¹, most_reliable=false)
 (Fe2+, coordination=VI, econf=3d6, spin=LS, crystal_radius=75.0 pm, ionic_radius=61.0 pm, ionic_potential=0.03279 e pm⁻¹, origin=estimated, , most_reliable=false)

julia> feir(;most_reliable=true) # filtering
3-element Vector{Mendeleev.IonicRadius}:
 (Fe2+, coordination=VI, econf=3d6, spin=HS, crystal_radius=92.0 pm, ionic_radius=78.0 pm, ionic_potential=0.02564 e pm⁻¹, origin=from r^3 vs V plots, , most_reliable=true)
 (Fe3+, coordination=IV, econf=3d5, spin=HS, crystal_radius=63.0 pm, ionic_radius=49.0 pm, ionic_potential=0.06122 e pm⁻¹, most_reliable=true)
 (Fe3+, coordination=VI, econf=3d5, spin=HS, crystal_radius=78.5 pm, ionic_radius=64.5 pm, ionic_potential=0.04651 e pm⁻¹, origin=from r^3 vs V plots, , most_reliable=true)
 
julia> feir(;charge=2, coordination=:VI, econf="3d6", spin=:HS, most_reliable=true) #filtering
1-element Vector{Mendeleev.IonicRadius}:
 (Fe2+, coordination=VI, econf=3d6, spin=HS, crystal_radius=92.0 pm, ionic_radius=78.0 pm, ionic_potential=0.02564 e pm⁻¹, origin=from r^3 vs V plots, , most_reliable=true)

Mendeleev.IonicRadius — Type

IonicRadius

This struct describes ionic radius for a given charge state of a given element. It is not exported.

    atomic_number::Int
    charge::Int
    coordination::Symbol
    crystal_radius::typeof(1.0u"pm")
    econf::Union{String, Missing}
    ionic_radius::typeof(1.0u"pm")
    most_reliable::Union{Bool, Missing}
    origin::Union{String, Missing}
    spin::Union{Symbol, Missing}
    ionic_potential::typeof(1.0u"e_au"/u"pm")

Mendeleev.Isotope — Type

Isotope

This struct describes a single isotope. Only naturally occuring stable or almost stable isotopes are included. It is not exported.

    atomic_number::Int
    mass_number::Int
    mass::typeof(1.0*u"u")
    abundance::Float64

Mendeleev.Isotopes — Type

Isotopes

This struct is a container for isotopes of an element. It provides access only by position(s) in the array of the (stable or almost stable) isotopes of this elements. More comprehensive isotopes data is published in a separate package IsotopeTable.jl. It is not exported.

Mendeleev.LiXue — Type

LiXue

This struct is a container for Li-Xue scale electronegativities of different ionic states of an element. It provides access by position(s) in the array of LiXueDSet structs for the given element, as well as filtering according to (; charge, coordination, spin). It is not exported.

Examples

julia> felx = chem_elements.Fe.eneg.Li;

julia> felx[2]
(Fe2+, coordination=IVSQ, spin=HS, value=4.826 pm⁻¹)

julia> felx[[2,3]]
2-element Vector{Mendeleev.LiXueDSet}:
 (Fe2+, coordination=IVSQ, spin=HS, value=4.826 pm⁻¹)
 (Fe2+, coordination=VI, spin=HS, value=4.092 pm⁻¹)

 julia> felx(;charge=2, spin=:HS) # filtering
 4-element Vector{Mendeleev.LiXueDSet}:
  (Fe2+, coordination=IV, spin=HS, value=4.889 pm⁻¹)
  (Fe2+, coordination=IVSQ, spin=HS, value=4.826 pm⁻¹)
  (Fe2+, coordination=VI, spin=HS, value=4.092 pm⁻¹)
  (Fe2+, coordination=VIII, spin=HS, value=3.551 pm⁻¹))

  julia> felx(;charge=2, spin=:HS, coordination=:VI) # filtering
  1-element Vector{Mendeleev.LiXueDSet}:
   (Fe2+, coordination=VI, spin=HS, value=4.092 pm⁻¹)

Mendeleev.LiXueDSet — Type

LiXueDSet

This struct describes electronegativity by Li-Xue scale for a given ionic states of an element. It is not exported.

    atomic_number::Int
    charge::Int
    coordination::Symbol
    spin::Union{Symbol, Missing}
    value::Float64 # dimensionality to be clarified yet

Mendeleev.ScreenConst — Type

ScreenConst

This struct describes screening constant for a given orbital of a given element. It is not exported.

    atomic_number::Int
    shell::Int
    orb_type::Int
    screening::Float64

Mendeleev.ScreenConstants — Type

ScreenConstants

This struct is a container for screening constants of an element. See examples below for indexing access. It is not exported.

Examples

julia> chem_elements[:Mo].sconst[2, 1]
Mo 2s: 11.1232

julia> chem_elements[:Mo].sconst[2, 's']
Mo 2s: 11.1232

julia> chem_elements[:Mo].sconst[2, "s"]
Mo 2s: 11.1232

julia> chem_elements[:Mo].sconst["2s"]
Mo 2s: 11.1232